+++ desktop/cmp/contextmenu/index.ts
@@ -1,2 +1,1 @@
 export * from './ContextMenu';
-export * from './StoreContextMenu';
