+++ desktop/hooks/UseContextMenu.ts
@@ -5,8 +5,8 @@
  * Copyright © 2022 Extremely Heavy Industries Inc.
  */
 import {XH} from '@xh/hoist/core';
-import {contextMenu as contextMenuEl} from '@xh/hoist/desktop/cmp/contextmenu/ContextMenu';
-import {ContextMenu, ContextMenuItem} from '@xh/hoist/kit/blueprint';
+import {contextMenu, ContextMenuSpec} from '@xh/hoist/desktop/cmp/contextmenu/ContextMenu';
+import {ContextMenu} from '@xh/hoist/kit/blueprint';
 import {isArray, isFunction, isUndefined, isEmpty} from 'lodash';
 import {ReactElement} from 'react';
 import {cloneElement, isValidElement} from 'react';
@@ -16,18 +16,14 @@ import {cloneElement, isValidElement} from 'react';
  *
  * @param child - element to be given context menu support.  Must specify Component
  *      that takes react context menu event as a prop (e.g. boxes, panel, div, etc).
- * @param contextMenu -  Array of ContextMenuItems or a function that receives the triggering event
- *      and returns such an array.  If null, or the number of items is empty, no menu will be
- *      rendered, and the event will be consumed. A ContextMenu element may also be provided.
+ * @param contextMenu -  Context Menu to be shown.  If null, or the number of items is empty,
+ *      no menu will be rendered, and the event will be consumed.
  */
-export function useContextMenu(
-    child?: ReactElement,
-    contextMenu?: ContextMenuItem[]|((e: MouseEvent) => ContextMenuItem[])|ReactElement
-): ReactElement {
-    if (!child || isUndefined(contextMenu)) return child;
+export function useContextMenu(child?: ReactElement, spec?: ContextMenuSpec): ReactElement {
+    if (!child || isUndefined(spec)) return child;
 
     const onContextMenu = (e: MouseEvent) => {
-        let contextMenuOutput = contextMenu;
+        let contextMenuOutput: ContextMenuSpec = spec;
 
         // 0) Skip if already consumed, otherwise consume (Adapted from Blueprint 'ContextMenuTarget')
         if (e.defaultPrevented) return;
@@ -38,7 +34,7 @@ export function useContextMenu(
             contextMenuOutput = contextMenuOutput(e);
         }
         if (isArray(contextMenuOutput)) {
-            contextMenuOutput = !isEmpty(contextMenuOutput) ? contextMenuEl({menuItems: contextMenuOutput}) : null;
+            contextMenuOutput = !isEmpty(contextMenuOutput) ? contextMenu({menuItems: contextMenuOutput}) : null;
         }
         if (contextMenuOutput && !isValidElement(contextMenuOutput)) {
             console.error("Incorrect specification of 'contextMenu' arg in useContextMenu()");
