+++ mobile/cmp/menu/MenuButton.ts
@@ -4,10 +4,10 @@
  *
  * Copyright © 2022 Extremely Heavy Industries Inc.
  */
-import {hoistCmp, HoistModel, PlainObject, MenuItem, useLocalModel} from '@xh/hoist/core';
+import {hoistCmp, HoistModel, useLocalModel, MenuItemLike} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {button, ButtonProps} from '@xh/hoist/mobile/cmp/button';
-import {popover} from '@xh/hoist/mobile/cmp/popover';
+import {popover, PopoverProps} from '@xh/hoist/mobile/cmp/popover';
 import '@xh/hoist/mobile/register';
 import {observable, action, makeObservable} from '@xh/hoist/mobx';
 import {ReactNode} from 'react';
@@ -16,21 +16,21 @@ import {menu} from './impl/Menu';
 
 export interface MenuButtonProps extends Omit<ButtonProps, 'title'> {
 
-    /** Array of MenuItems or configs to create them */
-    menuItems?: (MenuItem|PlainObject)[];
+    /** Array of MenuItems or spacers. */
+    menuItems?: MenuItemLike[];
 
     /** Position of menu relative to button */
     menuPosition?: 'top-left'|'top'|'top-right'|'right-top'|'right'|'right-bottom'|'bottom-right'|
         'bottom'|'bottom-left'|'left-bottom'|'left'|'left-top'|'auto';
 
-    /** Optional title to display above the menu */
+    /** Optional title to display above the menu. */
     title?: ReactNode;
 
-    /** True to disable user interaction */
+    /** True to disable user interaction. */
     disabled?: boolean;
 
-    /** Props passed to the internal popover */
-    popoverProps?: PlainObject;
+    /** Props passed to the internal popover. */
+    popoverProps?: PopoverProps;
 }
