+++ mobile/cmp/menu/index.ts
@@ -6,4 +6,3 @@
  */
 
 export * from './MenuButton';
-export * from './MenuItem';
